Opdaterede fra int til short.

+++ Torsdagsaflevering d. 06.02.2025/TaskTwo/TaskTwo.pde
@@ -1,8 +1,8 @@
 String teacherName = "Tess Løvgreen";// Skriv den rette datatype og et passende variabelnavn foran denne værdi: String teacherName= "Tess Løvgreen"
-int aYear = 365;
+short aYear = 365; // Bruger short, da vi vil gemme en lille data - Kunne også bruge int
 String cph = "København";
 boolean trueBoolean = true;
-int firstMath = 60*24;
+short firstMath = 60*24; // Igen - Vi kunne bruge int, istedet for short
 float date = 22.12;
 String text = "Hej, jeg heddder "+teacherName;
 boolean falseBoolean = false;
